A much needed and proud refactor
