+++ src/app.rs
@@ -1,0 +1,73 @@
+pub mod appdata;
+mod creation;
+mod device;
+pub mod queue;
+mod swapchain;
+
+use crate::constants::*;
+use appdata::AppData;
+use creation::*;
+use device::pick_physical_device;
+use swapchain::create_swapchain;
+use swapchain::create_swapchain_image_views;
+
+use vulkanalia::loader::{LibloadingLoader, LIBRARY};
+use vulkanalia::prelude::v1_0::*;
+use vulkanalia::window as vk_window;
+
+use vulkanalia::vk::ExtDebugUtilsExtension;
+use vulkanalia::vk::KhrSurfaceExtension;
+use vulkanalia::vk::KhrSwapchainExtension;
+
+use winit::window::Window;
+
+use anyhow::{anyhow, Result};
+
+#[derive(Clone, Debug)]
+pub struct App {
+    entry: Entry,
+    instance: Instance,
+    data: AppData,
+    device: Device,
+}
+
+impl App {
+    pub unsafe fn create(window: &Window) -> Result<Self> {
+        let loader = LibloadingLoader::new(LIBRARY)?;
+        let entry = Entry::new(loader).map_err(|b| anyhow!("{}", b))?;
+        let mut data = AppData::default();
+        let instance = create_instance(window, &entry, &mut data)?;
+        data.surface = vk_window::create_surface(&instance, &window, &window)?;
+        pick_physical_device(&instance, &mut data)?;
+        let device = create_logical_device(&entry, &instance, &mut data)?;
+        create_swapchain(window, &instance, &device, &mut data)?;
+        create_swapchain_image_views(&device, &mut data)?;
+        Ok(Self {
+            entry,
+            instance,
+            data,
+            device,
+        })
+    }
+
+    pub unsafe fn render(&mut self, window: &Window) -> Result<()> {
+        Ok(())
+    }
+
+    pub unsafe fn destroy(&mut self) {
+        if VALIDATION_ENABLED {
+            self.instance
+                .destroy_debug_utils_messenger_ext(self.data.messenger, None)
+        }
+
+        self.device.destroy_device(None);
+        self.instance.destroy_surface_khr(self.data.surface, None);
+        self.instance.destroy_instance(None);
+
+        self.device.destroy_swapchain_khr(self.data.swapchain, None);
+        self.data
+            .swapchain_image_views
+            .iter()
+            .for_each(|v| self.device.destroy_image_view(*v, None));
+    }
+}
