+++ src/utils.rs
@@ -1,0 +1,5 @@
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+#[error("Missing {0}.")]
+pub struct SuitabilityError(pub &'static str);
